编译告警修复（符号整数和无符合整数比较）
# message auto-generated for no-merge-commit merge:
merge op_mr into master

编译告警修复（符号整数和无符合整数比较）

Created-by: msprofDeveloper

Match-id-e80d1ebb34d6bef716ad9d4b9a64b8307c6df8cd

+++ analysis/csrc/domain/services/parser/freq/freq_parser.cpp
@@ -81,7 +81,7 @@ uint32_t FreqParser::ParseData(DataInventory &dataInventory, const Infra::Contex
     }
     // 根据device启动时间过滤有效频率数据
     for (const auto& freqData : halUniData_) {
-        for (int i = 0; i < freqData.count; ++i) {
+        for (uint64_t i = 0; i < freqData.count; ++i) {
             auto freqLpmData = freqData.freqLpmDataS[i];
             if (freqLpmData.sysCnt < deviceStart.cntVct) {
                 lpmDataS[0].freq = freqLpmData.freq;
